update to 10864(2015-9-23)

+++ FL/src/os/win32/fl_dnd.cxx
@@ -334,7 +334,7 @@ public:
 			delete this;
 		return nTemp;
 	}
-	STDMETHODIMP GiveFeedback( ulong ) {
+	STDMETHODIMP GiveFeedback( DWORD ) {
 		return DRAGDROP_S_USEDEFAULTCURSORS;
 	}
 	STDMETHODIMP QueryContinueDrag( BOOL esc, DWORD keyState ) {
